+++ libpigpiodpp/include/pigpiodpp/gpiomode.hpp
@@ -3,14 +3,13 @@
 #ifdef HAVE_PIGPIO
 #include <pigpiod_if2.h>
 #else
-#include "pigpiod/pigpiodstubs.hpp"
+#include "pigpiodpp/pigpiodstubs.hpp"
 #endif
 
-namespace Lineside {
-  namespace PiGPIOd {
-    enum class GPIOMode {
-			 Input = PI_INPUT,
-			 Output = PI_OUTPUT
-    };
-  }
+namespace PiGPIOdpp {
+  //! Enumeration of GPIO modes (input or output)
+  enum class GPIOMode {
+		       Input = PI_INPUT,
+		       Output = PI_OUTPUT
+  };
 }
